requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# Search tool
duckduckgo-search>=6.0.0

# Google Gemini API
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
